fs: Explore the honeyfs directory for realfile names at init time.
The old method of checking the honeyfs directory whenever a file was
accessed for the first time required that the original path to the file
be known. If the file was renamed, copied, moved to a new directory, or
one of its parent directories was renamed before its first access, its
original path would be completely lost and the real filename would not be
resolved.

This new method ensures that all A_REALFILE attributes are populated
upfront, and the filesystem can be rearranged plenty without breaking
honeyfs.