Full docker support
Currently Docker images are build by a second git repository.
Changes to installation or starting cowrie would need to be done on
both. Merging this into one repository prevents that those will be
forgotten and makes it easier to understand why changes happen.

The project folder structure has been updated to better integrate with
docker and for better seperating code from other things like docs etc.
I choosed app for the the name but it can be anything. src would also be
a good option. Naming it cowrie on the other hand is, at least for me,
not a good choise because this would end up in local folders
`./cowrie/cowrie/cowrie`.

The dockerfile is a different one then the one from the docker-cworie
repository.
I choosed to use a python2-alpine linux. In the end this image has 55%
smaller image size than the Debian image. The build process is splitted
into to parts. The first image has everythign installed to compile the
python modules. The second one has only things installed which are
needed to run the daemon.
There is no need to install python-virtualenv because we are using
docker. We don't need that much layers.
Twisted can drop his privileges when starting the daemon when `--uid`
and `--gid`is passed. This works only with numerical id. The user
`nobody` is used for this. This is on Docker a good idea since there
should be only one service with this user running. In other systems
there might be several services using this daemon which is not a good
choise.
Using this option of twisted is only posible when the startup check if
running as root is not resulting in a `sys.exit()`. For now this is
disabled but there is no need to have this code at all. Unexperienced
users will be using the Docker image or following the install
instructions where a user is created.

Further plans from here could be:
- Travis automatically building new releases from master
- Travis using a second Docker image as "attacker" for testing code on
higher levels
- Adding a Dockerfile.dev for local development