Removing root process check on startup
We don't need to check for running the service as root.
Yes, it's a bad idea but there is no way to prevent a user from shooting
himself.
On Docker cowrie drops his privileges, on systemd it is described how to
install with a different user and for non systemd systems it is also
documented on how to install cowrie without privileges.