Full docker support
Currently Docker images are build by a second git repository.
Changes to installation or starting cowrie would need to be done on
both. Merging this into one repository prevents that those will be
forgotten and makes it easier to understand why changes happen.

The dockerfile is a different one then the one from the docker-cworie
repository.
I choosed to use a python2-alpine linux. In the end this image has 55%
smaller image size than the Debian image. The build process is splitted
into to parts. The first image has everythign installed to compile the
python modules. The second one has only things installed which are
needed to run the daemon.
There is no need to install python-virtualenv because we are using
docker. We don't need that much layers.
Twisted can drop his privileges when starting the daemon when `--uid`
and `--gid` is passed. This works only with numerical id. The user
nobody is used for this. This is on Docker a good idea since there
should be only one service with this user running. In other systems
there might be several services using this daemon which is not a good
choise.

When building a new Docker image for cowrie Docker multistage build
images are created running flake8 and unittests to ensure that all
future releases are stable and matching our code guidelines.

Bonus effect is when using this as a git pre-push-hook a developer
doesn't need to wait for travis to fail on an error.

Based on the current project structure we need a lot of `COPY`
instructions inside the dockerfile which has negative sideeffects.
- bloading the dockerfile up
- longer buildtimes
- more layers are created
- more diskspace is used

We should find a way to reduce this. Best way for doing this is keeping
the static files like `honeyfs` and `share` right next to the source
code.