Introducint ENV var COWRIE_DOCKER
The variable is used inside the docker image to let cowrie know that it
is running inside docker and don't need to perform the "running as root"
check.
Inside the docker image cowrie is started with the `--uid` and `--gid`
option and will drop to a different user then root.