Restructured Dockerfile, Added cowrie user
The image is now builded with a user and group for running in the later
image cowrie.

Also the build steps are re-aranged to save build time.
We assume that static files like `honeyfs` and `share` are less
frequently updated and can be build into the base image where every
other images is based on.