removed old code
I removed every code line which has been commented out and is not in
use. This might be a bit pitty for all the test cases but reduces the
overhead of noise inside the source code.

If ever code is needed we can use git to get it back so it's not lost.