Cleanup file headers
Copyright:
There was no rule for having a copyright in the top of the file or not.
Some had it but most of them didn't had it.
From my point it should be enough to have the copyright at the toplevel
of the project once and for all.
It's going to save a lot of space and we can better look for comments
inside our code.

Contributors:
Similar to the copyright. There was no policy for it. Some people
claimed their copyright and others didn't. The problem I see here is
that the copyright notes might get that old that the file header gets
wrong.
Basically we don't need any kind of file headers for this. We are using
git and anyone can lookup who wrote the code. And for those poeple who
want to claim something special or having their name inside the project
we can now use the CONTRIBUTORS.txt