Refactor UserDB config files
The data_path has changed to etc/. I'm not really happy with this but
I didn't had a better idea since we allow configs to be in ., etc/ and
/etc. Maybe we wanna changes this behaviour when we have a stable docker
release.

If the userdb.txt is not found Cowrie will load a default list.
The parser is now also a bit less error prone when parsing this file.